requirements.txt
# Core dependencies
torch>=1.12.0
torchvision>=0.13.0
torchaudio>=0.12.0

# Transformers and audio processing
transformers>=4.20.0
librosa>=0.9.0
soundfile>=0.10.0

# Computer vision and image processing
opencv-python>=4.6.0
Pillow>=9.0.0

# Scientific computing
numpy>=1.21.0
scipy>=1.8.0
pandas>=1.4.0

# Machine learning utilities
accelerate>=0.20.0
tqdm>=4.64.0
scikit-learn>=1.1.0

# Specialized libraries for diffusion models
einops>=0.6.0
torchdiffeq>=0.2.0
torchsde>=0.2.0
jsonmerge>=1.8.0

# Optional: for advanced features
# wandb>=0.13.0  # For experiment tracking
# tensorboard>=2.9.0  # For logging

# Development dependencies (optional)
# pytest>=7.0.0
# black>=22.0.0
# flake8>=4.0.0
